+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Dependencies (`@grafana/faro-core`): upgrade otel deps (#670).
 
 - Dependencies (`@grafana/faro-transport-otlp-http [experimental]`): upgrade otel deps (#670)
+
   - Note: some attributes have been changed due to otel semantic attributes spec:
     - `enduser.id` is replaced by `user.id`
     - `enduser.name` is replaced by `user.username`,